aws/terraform/provision_endpoint/src/scripts/requirements.txt
boto3~=1.26.138
ruamel.yaml~=0.17.26